requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dataclasses-json==0.6.7
frozenlist==1.7.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-google-genai==0.0.5
langgraph==0.0.15
langsmith==0.0.87
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.4
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.2
packaging==23.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.7.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==8.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
